Refactor Graph Canvas view into separate views to help performance, view recalculation and keep settings from resetting.
