+++ Fabric/Views/Graph/GraphBackground.swift
@@ -1,0 +1,18 @@
+//
+//  GraphBackground.swift
+//  Fabric
+//
+
+import SwiftUI
+
+struct GraphBackground: View
+{
+    let geom: GeometryProxy
+
+    var body: some View
+    {
+        Image("background")
+            .resizable(resizingMode: .tile)
+            .offset(-geom.size / 2)
+    }
+}
